backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==4.0.1
certifi==2026.2.25
click==8.3.1
colorama==0.4.6
distro==1.9.0
ecdsa==0.19.1
fastapi==0.135.1
google-auth==2.41.1
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.27.2
idna==3.11
gunicorn==21.2.0
jiter==0.13.0
openai==2.24.0
passlib==1.7.4
psycopg2-binary==2.9.11
pyasn1==0.6.2
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.2
python-jose==3.5.0
redis==7.2.1
rsa==4.9.1
six==1.17.0
groq==0.4.2
sniffio==1.3.1
SQLAlchemy==2.0.48
starlette==0.52.1
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.41.0
pydantic-settings==2.13.1
python-multipart==0.0.20
requests==2.32.5

